Raise IOError when TCPConnection.read(length) recieves no data.
This commit fixes #24 by simply checking whether the socket received any data at all. If it did not, an IOError is raised.
It might be that we should try to recieve multiple times before the IOError is raised, but this is not implemented in this commit as to avoid bloat.