Fix varint reading and writing. Closes #141
Make varint reading and writing functions use C int 32's with ctypes module.
Important: Writing converts value to unsigned_int32, so that signed bit get forced to turn into a part of the number for writing, sort of breaking the number so python handles it like Java and C would. Reading converts final value back to signed_int32 so that the number is correct when python tries to use it.